Split chat stylesheets into separate file
