+++ src/react-components/lobby-chat-box.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
-import styles from "../assets/stylesheets/ui-root.scss";
+import styles from "../assets/stylesheets/message-entry.scss";
 import { WithHoverSound } from "./wrap-with-audio";
 import { handleTextFieldFocus, handleTextFieldBlur } from "../utils/focus-utils";
 import { faPaperPlane } from "@fortawesome/free-solid-svg-icons/faPaperPlane";
